+++ client/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@material-ui/core": "^4.10.0",
+    "axios": "^0.19.2",
     "dayjs": "^1.8.28",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
