+++ client/src/components/Upload/style.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import Button from "../Button";
+import Button from "../common/Button";
 
 export const UploadWrapper = styled.div`
   max-width: 680px;
