+++ client/src/containers/ArticleContainer.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Layout from "../components/Layout";
+import Layout from "../components/common/Layout";
 import { useState } from "react";
 import Article from "../components/Article";
