--- client/src/lib/api/index.js
@@ -1,3 +1,0 @@
-const api = `http://www.daangn.site`;
-
-export default api;
