--- client/src/lib/api/product.js
@@ -1,3 +1,0 @@
-import api from "./index";
-
-export const getProduct = (id) => fetch(`${api}/products/${id}`);
