+++ client/src/pages/LoginPage.js
@@ -1,6 +1,6 @@
 import React from "react";
 import LoginContainer from "../containers/LoginContainer";
-import Layout from "../components/Layout";
+import Layout from "../components/common/Layout";
 
 const LoginPage = () => {
   return (
