+++ client/src/pages/OtherUserPage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Layout from "../components/Layout";
+import Layout from "../components/common/Layout";
 import OtherUserContainer from "../containers/OtherUserContainer";
 
 const OtherUserPage = () => {
