+++ client/src/pages/RegisterPage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Layout from "../components/Layout";
+import Layout from "../components/common/Layout";
 import RegisterContainer from "../containers/RegisterContainer";
 
 const RegisterPage = () => {
