+++ client/src/pages/UploadPage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Layout from "../components/Layout";
+import Layout from "../components/common/Layout";
 import UploadContainer from "../containers/UploadContainer";
 
 const UploadPage = () => {
